requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
psycopg[binary]==3.2.13
PyJWT==2.10.1
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
httpx==0.28.1
pytest==8.3.5
pytest-asyncio==0.26.0
